+++ apps/docs/[doc].paths.ts
@@ -1,10 +1,9 @@
 import type { Feature } from '@readme-widget-hub/render'
 import type { WidgetTree } from '@readme-widget-hub/widget'
 import path from 'node:path'
-import { Manager } from '@readme-widget-hub/manager'
 import { renderIndexDoc, renderWidgetDoc } from '@readme-widget-hub/render'
 import { path2Url } from '@readme-widget-hub/utils'
-import { loadEnv } from './.vitepress/env'
+import { env, locales, manager, rootDir } from './manager'
 
 interface Path {
   params: {
@@ -12,15 +11,6 @@ interface Path {
   }
   content: string
 }
-
-const rootDir = path.resolve(__dirname, '../../')
-const env = loadEnv(rootDir)
-const manager = new Manager({
-  defaultLocaleCode: env.VITE_DEFAULT_LOCALE_CODE,
-  absWidgetsDir: path.join(rootDir, env.VITE_WIDGETS_DIR),
-  absMetaFilePath: path.join(rootDir, env.VITE_META_FILE_PATH),
-})
-const locales = manager.getLocales()
 
 function widgetFilePath2Url(localeUrlPrefix: string, filePath: string): string {
   return path2Url(path.join(localeUrlPrefix, path.relative(rootDir, filePath))).slice(0, -5)